requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
fastapi==0.95.1
h11==0.14.0
idna==3.4
numpy==1.24.3
opencv-python==4.7.0.72
Pillow==9.5.0
pydantic==1.10.7
requests==2.28.2
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
jose==1.0.0
jwt==1.3.1
bcrypt==4.0.1
passlib==1.7.4
bson==0.5.10
email_validator==2.0.0
motor==3.2.0
pymongo==4.4.1
crypto==1.4.1
pycryptodome==3.18.0
python-jose==3.3.0
python-multipart==0.0.6